backend/requirements.txt
fashion-clip>=0.2.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
transformers>=4.30.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
